+++ docs/release/release_notes.md
@@ -1,5 +1,5 @@
 ## Release 1.0
 This is the first release of the Bitpay extension.
 
-* Supports Bitpay via contribution pages.
+* Supports Bitpay payments via contribution pages.